backend/requirements.txt
MetaTrader5>=5.0.5572
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
pydantic==2.9.2
